chore: migrate storybook stories to tsx
