+++ src/components/StartExploringPage.stories.tsx
@@ -1,11 +1,9 @@
 import React from 'react'
 import i18n from '../i18n-decorator'
 import StartExploringPage from './StartExploringPage'
+import { Meta } from '@storybook/react'
 
-/**
- * @type {import('@storybook/react').Meta<typeof StartExploringPage>}
- */
-const meta = {
+const meta: Meta<typeof StartExploringPage> = {
   title: 'Start Exploring page',
   component: StartExploringPage,
   decorators: [i18n],
