+++ src/components/cid/Cid.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import Cid from './Cid'
+import { CID } from 'multiformats/cid'
 
 /**
  * @type {import('@storybook/react').Meta<typeof Cid>}
@@ -7,9 +8,9 @@ import Cid from './Cid'
 const meta = {
   title: 'CID',
   component: Cid,
-  render: () => (
-    <Cid />
-  )
+  // render: () => (
+  //   <Cid />
+  // )
 }
 
 export default meta
@@ -17,7 +18,7 @@ export default meta
 export const CidV0 = () => (
     <Cid
         className="db ma2 monospace"
-        value="QmYPNmahJAvkMTU6tDx5zvhEkoLzEFeTDz6azDCSNqzKkW"
+        value={CID.parse('QmYPNmahJAvkMTU6tDx5zvhEkoLzEFeTDz6azDCSNqzKkW')}
     />
 )
 
@@ -28,7 +29,7 @@ CidV0.story = {
 export const CidV1 = () => (
     <Cid
         className="db ma2 monospace"
-        value="zb2rhZMC2PFynWT7oBj7e6BpDpzge367etSQi6ZUA81EVVCxG"
+        value={CID.parse("zb2rhZMC2PFynWT7oBj7e6BpDpzge367etSQi6ZUA81EVVCxG")}
     />
 )
 
@@ -39,7 +40,7 @@ CidV1.story = {
 export const CidV1Sha3 = () => (
     <Cid
         className="db ma2 monospace"
-        value="zB7NbGN5wyfSbNNNwo3smZczHZutiWERdvWuMcHXTj393RnbhwsHjrP7bPDRPA79YWPbS69cZLWXSANcwUMmk4Rp3hP9Y"
+        value={CID.parse("zB7NbGN5wyfSbNNNwo3smZczHZutiWERdvWuMcHXTj393RnbhwsHjrP7bPDRPA79YWPbS69cZLWXSANcwUMmk4Rp3hP9Y")}
     />
 )
